Make a intro quest chain to set up the story. Revamp the treasurehunt quest. Add theme cards. Use card version of buttons and seperators.

+++ core/src/com/lyeeedar/Board/CompletionCondition/CompletionConditionDie.kt
@@ -18,7 +18,7 @@ class CompletionConditionDie : AbstractCompletionCondition()
 
 	override fun attachHandlers(grid: Grid)
 	{
-		maxHP = grid.level.player.getStat(Statistic.HEALTH)
+		maxHP = grid.level.player.getStat(Statistic.HEALTH).toInt()
 		hp = maxHP
 
 		grid.onAttacked += fun(c): Boolean {
